♻️Refactor. 랜덤 차트 데이터 생성 기능 모듈화 #60
