+++ view/src/pages/Home/index.tsx
@@ -3,35 +3,7 @@ import Chart from '../../entities/Chart';
 import Header from '../../widgets/Header';
 import Layout from '../../widgets/Layout';
 import candleMockData from './consts/candleData';
-
-const generateCandleData = (
-  existingData: {
-    date: Date;
-    open: number;
-    high: number;
-    low: number;
-    close: number;
-    volume: number;
-  }[]
-) => {
-  const lastDate = existingData[existingData.length - 1].date;
-  const newData = [];
-
-  for (let i = 1; i <= 5; i++) {
-    const date = new Date(lastDate);
-    date.setDate(date.getDate() + i); // 날짜를 하루씩 증가시킴
-
-    const open = Math.floor(Math.random() * 150) + 100;
-    const high = open + Math.floor(Math.random() * 20) + 5;
-    const low = open - Math.floor(Math.random() * 20) - 5;
-    const close = open + Math.floor(Math.random() * (high - low)) - (high - low) / 2;
-    const volume = Math.floor(Math.random() * 2000) + 1000;
-
-    newData.push({ date, open, high, low, close, volume });
-  }
-
-  return [...existingData, ...newData]; // 기존 데이터에 새 데이터를 추가하여 반환
-};
+import generateCandleData from './lib/generateCandleData';
 
 const Home = () => {
   const [candleData, setCandleData] = useState(candleMockData);
@@ -42,11 +14,9 @@ const Home = () => {
   return (
     <div>
       <Header />
-      <Layout paddingX="px-[22vw]">
-        <div>
-          <Chart data={candleData} />
-          <button onClick={handleButtonClick}>Generate More Data</button>
-        </div>
+      <Layout paddingX="px-[22vw]" flex={false}>
+        <Chart data={candleData} />
+        <button onClick={handleButtonClick}>Generate More Data</button>
       </Layout>
     </div>
   );
